chore(testing): Configure SonarQube to include TypeScript test files (#36)
* chore(testing): Configure SonarQube to include TypeScript test files

Added a configuration to include TypeScript test files located in the `src` directory for SonarQube analysis. This change helps in capturing and analyzing test-specific metrics to enhance code quality assessments.

* Integrate SonarCloud config into GitHub Workflow

Remove the standalone sonar-project.properties file and incorporate its configuration directly into the GitHub Actions workflow. This change centralizes the SonarCloud configuration, streamlining the maintenance and execution process within the CI/CD pipeline.

* Update ESLint workflow to adjust report file paths

Modified the ESLint run command to process the report by removing the workspace path prefix in the 'eslint-report.json'. This change ensures that file paths in the report are relative, thus simplifying file path management and making reports easier to manage and interpret.

* Update ESLint report path in GitHub Actions

This change modifies the path replacement in the ESLint report generation step to use '/github/workspace/' instead of removing the GitHub workspace path prefix. This ensures that paths in the ESLint report correctly reflect the directory structure used during the GitHub Action's execution. It aims to improve the accuracy and usefulness of the ESLint reports in CI workflows.

* Add SonarCloud configuration and streamline CI setup

Introduce a new `sonar-project.properties` file to configure SonarCloud integration for the project. This change centralizes the SonarCloud settings and removes redundant command-line parameters from the GitHub Actions workflow. Additionally, it fixes the ESLint report path in CI by stripping the workspace prefix.

* Refactor sonar configuration and update ESLint command path.

Reordered the paths for JavaScript and ESLint reports in sonar-project.properties to enhance readability and maintain consistency. Updated the ESLint command in package.json to specify the src/ directory, ensuring linting is applied only to source files.